+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## x.y.z (not yet released)
+
+  - Support instantiating arrays of instantiable types from database arrays.
+
 ## 0.7.1 (2015-03-27)
 
   - Delegate `PreparedStatement`-binding to `Dialect`s.